+++ service/api/tests/acceptance/api_acceptance_test.go
@@ -506,6 +506,7 @@ var _ = Describe("API Acceptance Tests", func() {
 	})
 
 	// TODO SMS Media tests
+	// TODO Incoming Phone Number tests
 	// TODO Queue Members tests
 	// TODO Conference tests
 	// TODO Conference Participants tests