+++ service/api/v2010/account/incoming_phone_number/api_op_update.go
@@ -1,0 +1,110 @@
+// Package incoming_phone_number contains auto-generated files. DO NOT MODIFY
+package incoming_phone_number
+
+import (
+	"context"
+	"net/http"
+
+	"github.com/RJPearson94/twilio-sdk-go/client"
+	"github.com/RJPearson94/twilio-sdk-go/utils"
+)
+
+// UpdateIncomingPhoneNumberInput defines input fields for updating a phone number
+type UpdateIncomingPhoneNumberInput struct {
+	APIVersion           *string `form:"ApiVersion,omitempty"`
+	AddressSid           *string `form:"AddressSid,omitempty"`
+	BundleSid            *string `form:"BundleSid,omitempty"`
+	EmergencyAddressSid  *string `form:"EmergencyAddressSid,omitempty"`
+	EmergencyStatus      *string `form:"EmergencyStatus,omitempty"`
+	FriendlyName         *string `form:"FriendlyName,omitempty"`
+	IdentitySid          *string `form:"IdentitySid,omitempty"`
+	SmsApplicationSid    *string `form:"SmsApplicationSid,omitempty"`
+	SmsFallbackMethod    *string `form:"SmsFallbackMethod,omitempty"`
+	SmsFallbackURL       *string `form:"SmsFallbackUrl,omitempty"`
+	SmsMethod            *string `form:"SmsMethod,omitempty"`
+	SmsURL               *string `form:"SmsUrl,omitempty"`
+	StatusCallback       *string `form:"StatusCallback,omitempty"`
+	StatusCallbackMethod *string `form:"StatusCallbackMethod,omitempty"`
+	TrunkSid             *string `form:"TrunkSid,omitempty"`
+	VoiceApplicationSid  *string `form:"VoiceApplicationSid,omitempty"`
+	VoiceCallerIDLookup  *bool   `form:"VoiceCallerIdLookup,omitempty"`
+	VoiceFallbackMethod  *string `form:"VoiceFallbackMethod,omitempty"`
+	VoiceFallbackURL     *string `form:"VoiceFallbackUrl,omitempty"`
+	VoiceMethod          *string `form:"VoiceMethod,omitempty"`
+	VoiceReceiveMode     *string `form:"VoiceReceiveMode,omitempty"`
+	VoiceURL             *string `form:"VoiceUrl,omitempty"`
+}
+
+type UpdateIncomingPhoneNumberResponseCapabilities struct {
+	Fax   bool `json:"fax"`
+	Mms   bool `json:"MMS"`
+	Sms   bool `json:"SMS"`
+	Voice bool `json:"voice"`
+}
+
+// UpdateIncomingPhoneNumberResponse defines the response fields for the updated phone number
+type UpdateIncomingPhoneNumberResponse struct {
+	APIVersion           string                                        `json:"api_version"`
+	AccountSid           string                                        `json:"account_sid"`
+	AddressRequirements  string                                        `json:"address_requirements"`
+	AddressSid           *string                                       `json:"address_sid,omitempty"`
+	Beta                 bool                                          `json:"beta"`
+	BundleSid            *string                                       `json:"bundle_sid,omitempty"`
+	Capabilities         UpdateIncomingPhoneNumberResponseCapabilities `json:"capabilities"`
+	DateCreated          utils.RFC2822Time                             `json:"date_created"`
+	DateUpdated          *utils.RFC2822Time                            `json:"date_updated,omitempty"`
+	EmergencyAddressSid  *string                                       `json:"emergency_address_sid,omitempty"`
+	EmergencyStatus      string                                        `json:"emergency_status"`
+	FriendlyName         *string                                       `json:"friendly_name,omitempty"`
+	IdentitySid          *string                                       `json:"identity_sid,omitempty"`
+	Origin               string                                        `json:"origin"`
+	PhoneNumber          string                                        `json:"phone_number"`
+	Sid                  string                                        `json:"sid"`
+	SmsApplicationSid    *string                                       `json:"sms_application_sid,omitempty"`
+	SmsFallbackMethod    string                                        `json:"sms_fallback_method"`
+	SmsFallbackURL       *string                                       `json:"sms_fallback_url,omitempty"`
+	SmsMethod            string                                        `json:"sms_method"`
+	SmsURL               *string                                       `json:"sms_url,omitempty"`
+	Status               string                                        `json:"status"`
+	StatusCallback       *string                                       `json:"status_callback,omitempty"`
+	StatusCallbackMethod string                                        `json:"status_callback_method"`
+	TrunkSid             *string                                       `json:"trunk_sid,omitempty"`
+	VoiceApplicationSid  *string                                       `json:"voice_application_sid,omitempty"`
+	VoiceCallerIDLookup  bool                                          `json:"voice_caller_id_lookup"`
+	VoiceFallbackMethod  string                                        `json:"voice_fallback_method"`
+	VoiceFallbackURL     *string                                       `json:"voice_fallback_url,omitempty"`
+	VoiceMethod          string                                        `json:"voice_method"`
+	VoiceReceiveMode     string                                        `json:"voice_receive_mode"`
+	VoiceURL             *string                                       `json:"voice_url,omitempty"`
+}
+
+// Update modifies a phone number resource
+// See https://www.twilio.com/docs/phone-numbers/api/incomingphonenumber-resource#update-an-incomingphonenumber-resource for more details
+// Context is defaulted to Background. See https://golang.org/pkg/context/#Background for more information
+func (c Client) Update(input *UpdateIncomingPhoneNumberInput) (*UpdateIncomingPhoneNumberResponse, error) {
+	return c.UpdateWithContext(context.Background(), input)
+}
+
+// UpdateWithContext modifies a phone number resource
+// See https://www.twilio.com/docs/phone-numbers/api/incomingphonenumber-resource#update-an-incomingphonenumber-resource for more details
+func (c Client) UpdateWithContext(context context.Context, input *UpdateIncomingPhoneNumberInput) (*UpdateIncomingPhoneNumberResponse, error) {
+	op := client.Operation{
+		Method:      http.MethodPost,
+		URI:         "/Accounts/{accountSid}/IncomingPhoneNumbers/{sid}.json",
+		ContentType: client.URLEncoded,
+		PathParams: map[string]string{
+			"accountSid": c.accountSid,
+			"sid":        c.sid,
+		},
+	}
+
+	if input == nil {
+		input = &UpdateIncomingPhoneNumberInput{}
+	}
+
+	response := &UpdateIncomingPhoneNumberResponse{}
+	if err := c.client.Send(context, op, input, response); err != nil {
+		return nil, err
+	}
+	return response, nil
+}